Changes according to thre review
