--- src/components/pages/Executors/ExecutorDetails/ExecutorSettings/Definition/index.ts
@@ -1,1 +1,0 @@
-export {default} from './Definition';
