+++ src/components/pages/Executors/ExecutorDetails/ExecutorSettings/ExecutorDefinition/index.ts
@@ -1,0 +1,1 @@
+export {default} from './ExecutorDefinition';
