+++ src/components/pages/Sources/SourceDetails/SourceSettings/Definition/index.ts
@@ -1,1 +1,1 @@
-export {default} from './Definition';
+export {default} from './SourceSettingsDefinition';
